Add a way to do local development by using `startdev.sh`, also move checkEnv to start scripts instead

+++ startdev.sh
@@ -3,6 +3,8 @@
 RC='\033[0m'
 RED='\033[0;31m'
 
+ISLOCAL=false
+
 # Function to fetch the latest release tag from the GitHub API
 get_latest_release() {
   local latest_release
@@ -28,6 +30,62 @@ redirect_to_latest_pre_release() {
   echo "Using URL: $url"  # Log the URL being used
 }
 
+PACKAGER=""
+DTYPE=""
+
+command_exists() {
+    which "$1" >/dev/null 2>&1
+}
+
+checkEnv() {
+    ## Check for requirements.
+    REQUIREMENTS='curl groups sudo'
+    for req in $REQUIREMENTS; do
+        if ! command_exists "$req"; then
+            printf "${RED}To run me, you need: %s${RC}\n" "$REQUIREMENTS"
+            exit 1
+        fi
+    done
+
+    ## Check Package Handler
+    PACKAGEMANAGER='apt-get dnf pacman zypper'
+    for pgm in $PACKAGEMANAGER; do
+        if command_exists "$pgm"; then
+            PACKAGER="$pgm"
+            printf "Using %s\n" "$pgm"
+            break
+        fi
+    done
+
+    if [ -z "$PACKAGER" ]; then
+        printf "${RED}Can't find a supported package manager${RC}\n"
+        exit 1
+    fi
+
+    ## Check SuperUser Group
+    SUPERUSERGROUP='wheel sudo root'
+    for sug in $SUPERUSERGROUP; do
+        if groups | grep -q "$sug"; then
+            SUGROUP="$sug"
+            printf "Super user group %s\n" "$SUGROUP"
+            break
+        fi
+    done
+
+    ## Check if member of the sudo group.
+    if ! groups | grep -q "$SUGROUP"; then
+        printf "${RED}You need to be a member of the sudo group to run me!${RC}\n"
+        exit 1
+    fi
+
+    DTYPE="unknown"  # Default to unknown
+    # Use /etc/os-release for modern distro identification
+    if [ -f /etc/os-release ]; then
+        . /etc/os-release
+        DTYPE="$ID"
+    fi
+}
+
 check() {
     local exit_code=$1
     local message=$2
@@ -38,20 +96,28 @@ check() {
     fi
 }
 
-redirect_to_latest_pre_release
+checkEnv
 
-TMPFILE=$(mktemp)
-check $? "Creating the temporary file"
+if [[ "$ISLOCAL" == "true" ]]; then
+    PKGR="$PACKAGER" DT="$DTYPE" cargo run
+else
+    echo "oh no"
+    exit 0
+    redirect_to_latest_pre_release
 
-echo "Downloading linutil from $url"  # Log the download attempt
-curl -fsL $url -o $TMPFILE
-check $? "Downloading linutil"
+    TMPFILE=$(mktemp)
+    check $? "Creating the temporary file"
 
-chmod +x $TMPFILE
-check $? "Making linutil executable"
+    echo "Downloading linutil from $url"  # Log the download attempt
+    curl -fsL $url -o $TMPFILE
+    check $? "Downloading linutil"
 
-"$TMPFILE"
-check $? "Executing linutil"
+    chmod +x $TMPFILE
+    check $? "Making linutil executable"
 
-rm -f $TMPFILE
-check $? "Deleting the temporary file"
+    PKGR="$PACKAGER-start" DT="$DTYPE-start" "$TMPFILE"
+    check $? "Executing linutil"
+
+    rm -f $TMPFILE
+    check $? "Deleting the temporary file"
+fi
